Move root level components into ./components and ./layout
