+++ src/ui/components/app-icon.svelte
