+++ src/ui/components/back-button.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import IconArrow from "./components/icons/icon-arrow.svelte";
-  import Localized from "./localization/localized.svelte";
-  import { LocalizationKeys } from "./localization/supportedLanguages";
+  import IconArrow from "./icons/icon-arrow.svelte";
+  import Localized from "../localization/localized.svelte";
+  import { LocalizationKeys } from "../localization/supportedLanguages";
 
   export let disabled = false;
 </script>
