+++ src/ui/components/badge.svelte
