+++ src/ui/components/branding-info.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import AppIcon from "./app-icon.svelte";
-  import { type BrandingInfoResponse } from "../networking/responses/branding-response";
-  import { buildAssetURL } from "../networking/assets";
+  import { type BrandingInfoResponse } from "../../networking/responses/branding-response";
+  import { buildAssetURL } from "../../networking/assets";
 
   export let brandingInfo: BrandingInfoResponse | null = null;
