+++ src/ui/components/icons/icon-cart.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Icon from "../../assets/cart.svg?raw"; // Load SVG as a raw string
+  import Icon from "../../../assets/cart.svg?raw"; // Load SVG as a raw string
 </script>
 
 <div style="color:var(--rc-color-gray-text-dark);">
