+++ src/ui/components/icons/icon-error.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Icon from "../../assets/error.svg?raw"; // Load SVG as a raw string
+  import Icon from "../../../assets/error.svg?raw"; // Load SVG as a raw string
 </script>
 
 <div style="color:var(--rc-color-primary);" class="rcb-ui-asset-icon">
