+++ src/ui/components/icons/icon-lock.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Icon from "../../assets/lock.svg?raw"; // Load SVG as a raw string
+  import Icon from "../../../assets/lock.svg?raw"; // Load SVG as a raw string
 </script>
 
 <div style="color:var(--rc-color-accent);" class="rcb-ui-asset-icon">
