+++ src/ui/components/icons/icon-success.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Icon from "../../assets/success.svg?raw"; // Load SVG as a raw string
+  import Icon from "../../../assets/success.svg?raw"; // Load SVG as a raw string
 </script>
 
 <div style="color:var(--rc-color-primary);" class="rcb-ui-asset-icon">
