+++ src/ui/components/processing-animation.svelte
