+++ src/ui/rcb-ui-inner.svelte
@@ -2,7 +2,7 @@
   import Container from "./layout/container.svelte";
   import Layout from "./layout/layout.svelte";
   import NavBar from "./layout/navbar.svelte";
-  import SandboxBanner from "./sandbox-banner.svelte";
+  import SandboxBanner from "./components/sandbox-banner.svelte";
   import Main from "./layout/main-block.svelte";
   import StateNeedsPaymentInfo from "./states/state-needs-payment-info.svelte";
   import StateNeedsAuthInfo from "./states/state-needs-auth-info.svelte";
@@ -13,7 +13,7 @@
   import { type BrandingInfoResponse } from "../networking/responses/branding-response";
   import type { Product, PurchaseOption } from "../main";
   import ProductInfo from "./components/product-info.svelte";
-  import BrandingInfoUI from "./branding-info-ui.svelte";
+  import BrandingInfoUI from "./components/branding-info.svelte";
   import {
     PurchaseFlowError,
     PurchaseFlowErrorCode,
