+++ src/ui/states/state-needs-auth-info.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import Button from "../button.svelte";
-  import ModalFooter from "../modal-footer.svelte";
-  import ModalSection from "../modal-section.svelte";
+  import Button from "../components/button.svelte";
+  import ModalFooter from "../layout/modal-footer.svelte";
+  import ModalSection from "../layout/modal-section.svelte";
   import RowLayout from "../layout/row-layout.svelte";
   import { validateEmail } from "../../helpers/validators";
   import { PurchaseFlowError } from "../../helpers/purchase-operation-helper";
@@ -11,7 +11,7 @@
   import { Translator } from "../localization/translator";
 
   import { LocalizationKeys } from "../localization/supportedLanguages";
-  import SecureCheckoutRc from "../secure-checkout-rc.svelte";
+  import SecureCheckoutRc from "../components/secure-checkout-rc.svelte";
   import { type ContinueHandlerParams } from "../ui-types";
   import { eventsTrackerContextKey } from "../constants";
   import { type IEventsTracker } from "../../behavioural-events/events-tracker";
